refactor(dx): tracing instrumentation, CheckFlags struct, VMA walker abstraction
Maintainability and debuggability improvements across the workspace:

- Add `tracing` + `tracing-subscriber` to workspace deps; initialize subscriber
  in `main()` (RUST_LOG=debug for structured span output); apply `#[instrument(skip_all)]`
  to all 9 cmd_* dispatch functions
- Extract `CheckFlags` struct (#[derive(Default)]) to replace 12 positional bools
  in cmd_check signature and all 4 call sites — eliminates transposition bugs
- Add `make_reader(isf: &IsfBuilder)` helper to `memf-core::test_builders` as
  shared test infrastructure for walker tests
- Add `memf-linux::vma_walker` module with `for_each_task_vma` and `VmaEntry` —
  mirrors Windows `for_each_heap_region`; eliminates repeated VMA linked-list
  traversal boilerplate from 6+ walkers; 4 TDD tests included (3418 total pass)

+++ crates/memf-windows/src/debug_registers.rs
@@ -620,7 +620,7 @@ mod tests {
         // eproc page
         let mut eproc_page = vec![0u8; 0x1000];
         // _KPROCESS.DirectoryTableBase at +0x28
-        eproc_page[0x28..0x30].copy_from_slice(&(eproc_paddr as u64).to_le_bytes());
+        eproc_page[0x28..0x30].copy_from_slice(&eproc_paddr.to_le_bytes());
         // _KPROCESS.ThreadListHead.Flink at +0x30 → kthread+0x2F8
         eproc_page[0x30..0x38].copy_from_slice(&kthread_entry_vaddr.to_le_bytes());
         eproc_page[0x38..0x40].copy_from_slice(&kthread_entry_vaddr.to_le_bytes());
